Новая пошаговая игра #187 - Сделал базовую отрисовку кроссворда с масштабированием и перемещением
