+++ frontend/src/entities/cross-ward/lib/PanZoomContext.js
@@ -1,0 +1,3 @@
+import {createContext} from 'react';
+
+export const PanZoomContext = createContext(null);
