+++ frontend/src/entities/cross-ward/lib/hooks.js
@@ -1,0 +1,10 @@
+import {useContext} from 'react';
+import {PanZoomContext} from './PanZoomContext';
+
+export const usePanZoom = () => {
+  const context = useContext(PanZoomContext);
+  if (!context) {
+    throw new Error('usePanZoom must be used within PanZoomProvider');
+  }
+  return context;
+};
