+++ frontend/src/widgets/cross-ward/CrossWardContent.jsx
@@ -1,8 +1,9 @@
+import {PanZoomProvider} from '@entities/cross-ward/lib/PanZoomProvider';
+import {CrosswordCanvas} from '@entities/cross-ward/ui/CrosswardCanvas';
 import {useGame} from '@entities/game/lib/hooks';
 import {Box, CircularProgress} from '@mui/material';
 import {useSocket} from '@shared/lib/socket/hooks';
 import {useEffect} from 'react';
-import {CrosswordCanvas3} from '../../../trash/CrosswordCanvas3';
 
 export const CrossWardContent = () => {
   const {connected} = useSocket();
@@ -30,18 +31,9 @@ export const CrossWardContent = () => {
     );
   }
 
-  const matrix = [
-    ['К', 'О', 'Т', null, 'М', 'О', 'Р', 'Е'],
-    ['А', null, 'А', null, 'О', null, null, null],
-    ['Р', 'А', 'К', null, 'Р', 'Е', 'К', 'А'],
-    [null, null, null, null, 'О', null, null, null],
-    ['Л', 'Е', 'С', null, 'Н', 'О', 'С', ''],
-    ['О', null, 'Т', null, 'Ы', null, null, null],
-    ['С', 'Л', 'О', 'Н', 'К', 'А', null, null],
-  ];
-
   return (
-    // <CrosswordCanvas matrix={grid.reduce((acc, f) => {const rr = acc[f.y]; if (!rr) {acc[f.y] = []} acc[f.y][f.x] = f.letter; return acc}, [])}/>
-    <CrosswordCanvas3 matrix={grid}/>
+    <PanZoomProvider>
+      <CrosswordCanvas matrix={grid}/>
+    </PanZoomProvider>
   );
 };
